requirements.txt
aiofiles==23.1.0
Pillow==9.5.0